databas klar kanske

+++ Hospitalet/src/server/Database.java
@@ -1,19 +1,23 @@
 package server;
 
+import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
 
-public class Database {
+@SuppressWarnings("serial")
+public class Database implements Serializable {
 	private ArrayList<Journal> db;
 	private ArrayList<String> doctors;
 	private ArrayList<String> nurses;
 	private ArrayList<String> patients;
+	private Auditlog log;
 
 	public Database() {
 		db = new ArrayList<Journal>();
 		doctors = new ArrayList<String>();
 		nurses = new ArrayList<String>();
 		patients = new ArrayList<String>();
+		log = new Auditlog();
 
 		doctors.add("Alban Nwapa");
 		doctors.add("Jonny Bombay");
@@ -31,20 +35,36 @@ public class Database {
 		patients.add("Black Knight");
 	}
 
-	public void addJournal(Journal journal) {
-		db.add(journal);
-	}
-
-	public void deleteJournal(String journalID) {
-
-		for (Journal j : db) {
-			if (journalID.equals(j.getID())) {
-				db.remove(j);
-			}
+	public String addJournal(Journal journal, String name) {
+		if (doctors.contains(name)) {
+			db.add(journal);
+			log.addEntry(journal.getDoctor(), journal.getID(),
+					"Added new journal; Doctor = " + journal.getDoctor()
+							+ "; Nurse = " + journal.getNurse()
+							+ "; Patient = " + journal.getPatient());
+			return "Journal with ID " + journal.getID() + " added.";
 		}
-		
+		log.addEntry(name, journal.getID(),
+				"User tried to add jounral without permission.");
+		return "Journal not added. Permission denied!";
 	}
-	
+
+	public String deleteJournal(String journalID, String name) {
+		if (name.equals("StyrelsenMADDERFAKKER")) {
+			for (Journal j : db) {
+				if (journalID.equals(j.getID())) {
+					db.remove(j);
+
+				}
+			}
+			log.addEntry(name, journalID, "Journal deleted.");
+			return "Journal with ID " + journalID + " removed.";
+		}
+		log.addEntry(name, journalID,
+				"User tried to delete journal without permission");
+		return "Journal not deleted. Permission Denied!";
+	}
+
 	public Journal getJournal(String journalID, String name, String division) {
 		Journal temp = null;
 		for (Journal j : db) {
@@ -54,24 +74,36 @@ public class Database {
 		if (patients.contains(name)) {
 
 			if (name.equals(temp.getPatient())) {
+				log.addEntry(name, temp.getID(), "User read journal.");
 				return temp;
 			}
 
 		}
 
 		else if (nurses.contains(name)) {
-			if (name.equals(temp.getNurse()) || division.equals(temp.getDivision())) {
+			if (name.equals(temp.getNurse())
+					|| division.equals(temp.getDivision())) {
+				log.addEntry(name, temp.getID(), "User read journal.");
 				return temp;
 			}
 		}
 
 		else if (doctors.contains(name)) {
-			if (name.equals(temp.getDoctor()) || division.equals(temp.getDivision())) {
+			if (name.equals(temp.getDoctor())
+					|| division.equals(temp.getDivision())) {
+				log.addEntry(name, temp.getID(), "User read journal.");
 				return temp;
 			}
 		}
-
-		return null;
+		else if (name.equals("StyrelsenMADDERFAKKER")) {
+			log.addEntry(name, temp.getID(), "User read journal.");
+			return temp;
+		}
+		log.addEntry(name, journalID,
+				"User tried to access journal without permission.");
+		Journal fakeJournal = new Journal("", "", "", "", "");
+		fakeJournal.changeText("Permission Denied");
+		return fakeJournal;
 	}
 
 	public List<Journal> getAssociatedJournals(String name, String division) {
@@ -82,6 +114,8 @@ public class Database {
 					list.add(j);
 				}
 			}
+			log.addEntry(name, "List of journals", "User retrieved associated journals");
+			return list;
 		}
 
 		else if (nurses.contains(name)) {
@@ -91,6 +125,8 @@ public class Database {
 					list.add(j);
 				}
 			}
+			log.addEntry(name, "List of journals", "User retrieved associated journals");
+			return list;
 		}
 
 		else if (doctors.contains(name)) {
@@ -100,7 +136,42 @@ public class Database {
 					list.add(j);
 				}
 			}
+			log.addEntry(name, "List of journals", "User retrieved associated journals");
+			return list;
 		}
-		return list;
+		
+		else if (name.equals("StyrelsenMADDERFAKKER")) {
+			for (Journal j : db) {
+					list.add(j);
+			}
+			log.addEntry(name, "List of journals", "User retrieved associated journals");
+			return list;
+		}
+		log.addEntry(name, "List of journals", "User tried to retieve list of journals. Access denied!");
+		return null;
+	}
+	
+	public String appendToText(String name, String journalID, String text) {
+		Journal temp = null;
+		for (Journal j : db) {
+			if (journalID.equals(j.getID()))
+				temp = j;
+		}
+		if (nurses.contains(name)) {
+			if (name.equals(temp.getNurse())) {
+				temp.changeText(text);
+				log.addEntry(name, temp.getID(), "User added text: " + text);
+				return "Text added to journal";
+			}
+		}
+		else if (nurses.contains(name)) {
+			if (name.equals(temp.getDoctor())) {
+				temp.changeText(text);
+				log.addEntry(name, temp.getID(), "User added text: " + text);
+				return "Text added to journal";
+			}
+		}
+		log.addEntry(name, journalID, "User tried to add text \"" + text + "\" to journal without permission.");
+		return "Did not add text to journal. Permission denied!";
 	}
 }
